Prevent compile time warnings.

+++ Pulp/Common/Hashing/Fnva1Hash.h
@@ -68,12 +68,12 @@ namespace Hash
 		{
 			constexpr static inline Fnv1aVal Hash(char const*const pStr, const uint32_t val)
 			{
-				return !*pStr ? val : Hash(pStr + 1, static_cast<uint32_t>(*pStr ^ (val * prime) & 0xFFFFFFFF));
+				return !*pStr ? val : Hash(pStr + 1, static_cast<uint32_t>((*pStr ^ (val * static_cast<uint64_t>(prime))) & 0xFFFFFFFF));
 			}
 
 			constexpr static inline Fnv1aVal Hash(char const*const pStr, const size_t strLen, const uint32_t val)
 			{
-				return (strLen == 0) ? val : Hash(pStr + 1, strLen - 1, static_cast<uint32_t>(*pStr ^ (val * prime) & 0xFFFFFFFF));
+				return (strLen == 0) ? val : Hash(pStr + 1, strLen - 1, static_cast<uint32_t>((*pStr ^ (val * static_cast<uint64_t>(prime))) & 0xFFFFFFFF));
 			}
 		} // namespace Internal
